requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Interpretability
shap>=0.43.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Medical Imaging
Pillow>=10.0.0
pydicom>=2.3.0

# Signal Processing
wfdb>=4.0.0        # PhysioNet ECG data reading

# Configuration & Utilities
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.13.0

# Notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: Federated Learning
# flwr>=1.5.0

# Optional: Model Export
# onnx>=1.14.0
# onnxruntime>=1.15.0
